Add UI to display the list of organizations
This commit adds to the app, the much needed UI which would show  the user a list of all organizations of which he is a part.
